feat: added reddit sentiment anaysis
